Fixed warnings

+++ Esercizio 1/src/merge_binary_insertion_sort_main.c
@@ -171,19 +171,19 @@ static void test_with_comparison_function(const char *file_path)
     printf("\nSort by int field...\n");
     clock_t start_time = clock();
     sorted_array_sort((void **)records, comparator_record_int_field, SIZE);
-    elapsed_time = (clock() - start_time) / 1000000.0;
+    elapsed_time = (double)(clock() - start_time) / 1000000.0;
     printf("Done in %f sec\n", elapsed_time);
 
     printf("\nSort by string field...\n");
     start_time = clock();
     sorted_array_sort((void **)records, comparator_record_string_field, SIZE);
-    elapsed_time = (clock() - start_time) / 1000000.0;
+    elapsed_time = (double)(clock() - start_time) / 1000000.0;
     printf("Done in %f sec\n", elapsed_time);
 
     printf("\nSort by double field...\n");
     start_time = clock();
     sorted_array_sort((void **)records, comparator_record_double_field, SIZE);
-    elapsed_time = (clock() - start_time) / 1000000.0;
+    elapsed_time = (double)(clock() - start_time) / 1000000.0;
     printf("Done in %f sec\n\n", elapsed_time);
 
     //print_records();
